+++ src/main/java/net/TheDgtl/Stargate/listeners/WorldEventsListener.java
@@ -1,0 +1,35 @@
+package net.TheDgtl.Stargate.listeners;
+
+import net.TheDgtl.Stargate.Portal;
+import net.TheDgtl.Stargate.Stargate;
+import org.bukkit.World;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.world.WorldLoadEvent;
+import org.bukkit.event.world.WorldUnloadEvent;
+
+public class WorldEventsListener implements Listener {
+    @EventHandler
+    public void onWorldLoad(WorldLoadEvent event) {
+        if (!Stargate.managedWorlds.contains(event.getWorld().getName())
+                && Portal.loadAllGates(event.getWorld())) {
+            Stargate.managedWorlds.add(event.getWorld().getName());
+        }
+    }
+
+    // We need to reload all gates on world unload, boo
+    @EventHandler
+    public void onWorldUnload(WorldUnloadEvent event) {
+        Stargate.debug("onWorldUnload", "Reloading all Stargates");
+        World w = event.getWorld();
+        if (Stargate.managedWorlds.contains(w.getName())) {
+            Stargate.managedWorlds.remove(w.getName());
+            Portal.clearGates();
+            for (World world : Stargate.server.getWorlds()) {
+                if (Stargate.managedWorlds.contains(world.getName())) {
+                    Portal.loadAllGates(world);
+                }
+            }
+        }
+    }
+}
